feat(examples):  * updated & cleaned up examples some more  * added features/feature flags to 'common' example boilerplate  * moved xpbd physics to common, behind a flag too  * updated xpbd example  * updated code for bevy_asset_loader changed api  * more cleanups
