+++ examples/bevy_gltf_blueprints/multiple_levels_multiple_blendfiles/src/game/in_game.rs
@@ -11,7 +11,6 @@ pub fn setup_game(
     models: Res<Assets<bevy::gltf::Gltf>>,
     mut next_game_state: ResMut<NextState<GameState>>,
 ) {
-    println!("setting up all stuff");
     commands.insert_resource(AmbientLight {
         color: Color::WHITE,
         brightness: 0.2,
