+++ examples/common/Cargo.toml
@@ -4,10 +4,18 @@ version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 
+[features]
+blueprints=["dep:bevy_gltf_blueprints"]
+physics_rapier = ["dep:bevy_rapier3d"]
+physics_xpbd = ["dep:bevy_xpbd_3d"]
+default = ["blueprints", "physics_rapier"]
+
+
 [dependencies]
 bevy="0.13"
-bevy_gltf_blueprints = { path = "../../crates/bevy_gltf_blueprints" }
-bevy_rapier3d = { version = "0.25.0", features = [ "serde-serialize", "debug-render-3d", "enhanced-determinism"] }
+bevy_gltf_blueprints = { path = "../../crates/bevy_gltf_blueprints", optional = true }
+bevy_rapier3d = { version = "0.25", features = [ "serde-serialize", "debug-render-3d", "enhanced-determinism"] , optional = true }
+bevy_xpbd_3d = { version = "0.4", optional = true }
 bevy_asset_loader = { version = "0.20", features = ["standard_dynamic_assets" ]} 
 #bevy_editor_pls = { version = "0.6" }
 rand = "0.8.5"
