+++ examples/common/src/assets/mod.rs
@@ -15,21 +15,21 @@ impl Plugin for AssetsPlugin {
         app
             // load core assets (ie assets needed in the main menu, and everywhere else before loading more assets in game)
             .add_loading_state(
-                LoadingState::new(AppState::CoreLoading).continue_to_state(AppState::MenuRunning),
+                LoadingState::new(AppState::CoreLoading)
+                    .continue_to_state(AppState::MenuRunning)
+                    .with_dynamic_assets_file::<StandardDynamicAssetCollection>(
+                        "assets_core.assets.ron",
+                    )
+                    .load_collection::<CoreAssets>()
             )
-            .add_dynamic_collection_to_loading_state::<_, StandardDynamicAssetCollection>(
-                AppState::CoreLoading,
-                "assets_core.assets.ron",
-            )
-            .add_collection_to_loading_state::<_, CoreAssets>(AppState::CoreLoading)
             // load game assets
             .add_loading_state(
-                LoadingState::new(AppState::AppLoading).continue_to_state(AppState::AppRunning),
-            )
-            .add_dynamic_collection_to_loading_state::<_, StandardDynamicAssetCollection>(
-                AppState::AppLoading,
-                "assets_game.assets.ron",
-            )
-            .add_collection_to_loading_state::<_, GameAssets>(AppState::AppLoading);
+                LoadingState::new(AppState::AppLoading)
+                    .continue_to_state(AppState::AppRunning)
+                    .with_dynamic_assets_file::<StandardDynamicAssetCollection>(
+                        "assets_game.assets.ron",
+                    )
+                    .load_collection::<GameAssets>()
+            );
     }
 }
