+++ examples/common/src/core/mod.rs
@@ -7,8 +7,16 @@ pub use lighting::*;
 //pub mod relationships;
 //pub use relationships::*;
 
-pub mod physics;
-pub use physics::*;
+#[cfg(feature = "physics_rapier")]
+pub mod physics_rapier;
+#[cfg(feature = "physics_rapier")]
+pub use physics_rapier::*;
+
+#[cfg(feature = "physics_xpbd")]
+pub mod physics_xpbd;
+#[cfg(feature = "physics_xpbd")]
+pub use physics_xpbd::*;
+
 
 use bevy::prelude::*;
