+++ examples/common/src/core/physics_rapier/physics_replace_proxies.rs
