+++ examples/common/src/core/physics_rapier/utils.rs
