+++ examples/common/src/core/physics_xpbd/controls.rs
@@ -1,5 +1,5 @@
 use bevy::ecs::system::Res;
-use bevy::gizmos::config::{GizmoConfig, GizmoConfigStore};
+use bevy::gizmos::config::GizmoConfigStore;
 use bevy::input::keyboard::KeyCode;
 use bevy::input::ButtonInput;
 use bevy::log::info;
@@ -25,14 +25,6 @@ pub fn toggle_physics_debug(
     if keycode.just_pressed(KeyCode::KeyD) {
         let config = config_store.config_mut::<PhysicsGizmos>().0;
         config.enabled = !config.enabled;
-        println!("BLAAA");
-        /* 
-        .insert_gizmo_group(
-            PhysicsGizmos {
-                aabb_color: Some(Color::WHITE),
-                ..default()
-            },
-            GizmoConfig::default(),
-        )*/
+        // config.aabb_color = Some(Color::WHITE);
     }
 }
