+++ examples/common/src/core/physics_xpbd/physics_replace_proxies.rs
