+++ examples/common/src/core/physics_xpbd/utils.rs
