+++ testing/bevy_registry_export/basic/src/game/in_game.rs
@@ -11,7 +11,6 @@ pub fn setup_game(
     models: Res<Assets<bevy::gltf::Gltf>>,
     mut next_game_state: ResMut<NextState<GameState>>,
 ) {
-    println!("setting up all stuff");
     commands.insert_resource(AmbientLight {
         color: Color::WHITE,
         brightness: 0.2,
@@ -41,12 +40,12 @@ pub fn setup_game(
 struct UnregisteredComponent;
 
 pub fn spawn_test(
-    keycode: Res<Input<KeyCode>>,
+    keycode: Res<ButtonInput<KeyCode>>,
     mut commands: Commands,
 
     mut game_world: Query<(Entity, &Children), With<GameWorldTag>>,
 ) {
-    if keycode.just_pressed(KeyCode::T) {
+    if keycode.just_pressed(KeyCode::KeyT) {
         let world = game_world.single_mut();
         let world = world.1[0];
 
@@ -77,50 +76,6 @@ pub fn spawn_test(
                     linvel: Vec3::new(vel_x, vel_y, vel_z),
                     angvel: Vec3::new(0.0, 0.0, 0.0),
                 },
-            ))
-            .id();
-        commands.entity(world).add_child(new_entity);
-    }
-}
-
-pub fn spawn_test_unregisted_components(
-    keycode: Res<Input<KeyCode>>,
-    mut commands: Commands,
-
-    mut game_world: Query<(Entity, &Children), With<GameWorldTag>>,
-) {
-    if keycode.just_pressed(KeyCode::U) {
-        let world = game_world.single_mut();
-        let world = world.1[0];
-
-        let mut rng = rand::thread_rng();
-        let range = 5.5;
-        let x: f32 = rng.gen_range(-range..range);
-        let y: f32 = rng.gen_range(-range..range);
-
-        let mut rng = rand::thread_rng();
-        let range = 0.8;
-        let vel_x: f32 = rng.gen_range(-range..range);
-        let vel_y: f32 = rng.gen_range(2.0..2.5);
-        let vel_z: f32 = rng.gen_range(-range..range);
-
-        let name_index: u64 = rng.gen();
-
-        let new_entity = commands
-            .spawn((
-                BluePrintBundle {
-                    blueprint: BlueprintName("Health_Pickup".to_string()),
-                    ..Default::default()
-                },
-                bevy::prelude::Name::from(format!("test{}", name_index)),
-                // BlueprintName("Health_Pickup".to_string()),
-                // SpawnHere,
-                TransformBundle::from_transform(Transform::from_xyz(x, 2.0, y)),
-                Velocity {
-                    linvel: Vec3::new(vel_x, vel_y, vel_z),
-                    angvel: Vec3::new(0.0, 0.0, 0.0),
-                },
-                UnregisteredComponent,
             ))
             .id();
         commands.entity(world).add_child(new_entity);
